Color Picker (HTML, CSS, JS)
This task required creating a color picker using HTML, CSS and JavaScript. The user could select a color and its transparency using two sliders. Within the square where the color was selected, a cursor allowed the user to click anywhere to choose a color, and the HEX code of the selected color is shown below.